+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_help.py
@@ -2374,6 +2374,9 @@ examples:
   - name: Create a basic internal load balancer on a specific virtual network and subnet.
     text: >
         az network lb create -g MyResourceGroup -n MyLb --sku Basic --vnet-name MyVnet --subnet MySubnet
+  - name: Create a basic load balancer on a subnet of a pre-existing virtual network.
+    text: >
+        az network lb create -g MyResourceGroup -n MyLb --sku Basic --subnet {subnetID}
   - name: Create a basic zone flavored internal load balancer, through provisioning a zonal public ip.
     text: >
         az network lb create -g MyResourceGroup -n MyLb --sku Basic --public-ip-zone 2