+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/commands.py
@@ -313,3 +313,6 @@ def load_command_table(self, _):
     with self.command_group('account management-group subscription', resource_managementgroups_subscriptions_sdk, client_factory=cf_management_group_subscriptions) as g:
         g.custom_command('add', 'cli_managementgroups_subscription_add')
         g.custom_command('remove', 'cli_managementgroups_subscription_remove')
+
+    with self.command_group('') as g:
+        g.custom_command('rest', 'rest_call', is_preview=True)