+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+* webapp, functionapp: Updating to the latest Python SDK version
+* functionapp: add slot support to functionapps
+
 0.2.20
 ++++++
 * functionapp: deprecated `az functionapp devops-build` command will be removed in next release